+++ pandokia/helpers/test_in_db.py
@@ -1,7 +1,7 @@
 import os
 import re
 import pandokia.helpers.dbaccess as dbaccess
-import pandokia.helpers.pycode as pycode
+import utils.helpers.pycode as pycode
 
 query = {
     'test_run': 'pdkrun_test',
