+++ pandokia/pcgi_summary.py
@@ -9,8 +9,8 @@ import copy
 import time
 from html import escape as html_escape
 
-import pandokia.lib as lib
-import pandokia.text_table as text_table
+import utils.lib as lib
+import utils.text_table as text_table
 
 try:
     from urllib.parse import quote
