+++ tests/tests_new/database/library/csv_t.py
@@ -1,4 +1,4 @@
-import pandokia.helpers.minipyt as minipyt
+import utils.helpers.minipyt as minipyt
 minipyt.noseguard()
 
 dbx = None  # will be assigned after we are imported
