+++ tests/tests_new/database/library/sqlite1.py
@@ -20,7 +20,7 @@ csv_t.dbx = dbx
 
 from csv_t import *
 
-import pandokia.helpers.minipyt as minipyt
+import utils.helpers.minipyt as minipyt
 
 
 @minipyt.test
