+++ tests/tests_new/helpers/compare_files.py
@@ -2,8 +2,8 @@ import os.path
 import json
 import time
 
-import pandokia.helpers.minipyt as mph
-import pandokia.helpers.pycode as pycode
+import utils.helpers.minipyt as mph
+import utils.helpers.pycode as pycode
 
 # software under test:
 import utils.helpers.filecomp as filecomp
