+++ tests/tests_new/pdkrun_test_data/helpers/json_helper.py
@@ -1,7 +1,7 @@
 # tests of the examples doc/source/helper_json.rst
 
 
-import pandokia.helpers.minipyt as mph
+import utils.helpers.minipyt as mph
 
 
 @mph.nottest
