+++ tests/tests_new/pdkrun_test_data/minipyt/class_1.py
@@ -1,4 +1,4 @@
-import pandokia.helpers.minipyt as mpt
+import utils.helpers.minipyt as mpt
 
 
 # notice that an object that does not set minipyt_shared will not get
