+++ tests/tests_new/pdkrun_test_data/minipyt/function.py
@@ -1,4 +1,4 @@
-import pandokia.helpers.minipyt as mpt
+import utils.helpers.minipyt as mpt
 import nose.tools as nt
 import time
