+++ tests/tests_new/pdkrun_test_data/minipyt/pycode.py
@@ -1,4 +1,4 @@
-import pandokia.helpers.pycode as pycode
+import utils.helpers.pycode as pycode
 
 with pycode.test('pass_test') as t:
     t.tda['yes'] = 'yes'
