+++ tests/tests_new/pdkrun_test_data/minipyt/tk122.py
@@ -1,4 +1,4 @@
-import pandokia.helpers.pycode as pycode
+import utils.helpers.pycode as pycode
 import pandokia.helpers.runner_minipyt as runner_minipyt
